MachineLearning

Description

Documentation

addBiasDimension :: Matrix -> Matrix

Add biad dimension to the future matrix

removeBiasDimension :: Matrix -> Matrix

Remove biad dimension

meanStddev :: Matrix R -> (Matrix R, Matrix R)

Caclulates mean and stddev values of every feature. Takes feature matrix X, returns pair of vectors of means and stddevs.

featureNormalization :: Fractional a => (a, a) -> a -> a

mapFeatures :: Int -> Matrix -> Matrix

Maps the features into all polynomial terms of X up to the degree-th power

splitToXY :: Element t => Matrix t -> (Matrix t, Vector t)

Splits data matrix to features matrix X and vector of outputs y