MachineLearning.Classification

Description

Binary and Multiclass Classification.

Documentation

data MinimizeMethod

Constructors

GradientDescent R
Gradient descent, takes alpha. Requires feature normalization.

ConjugateGradientFR R R
Fletcher-Reeves conjugate gradient algorithm, takes size of first trial step (0.1 is fine) and tol (0.1 is fine).

ConjugateGradientPR R R
Polak-Ribiere conjugate gradient algorithm. takes size of first trial step (0.1 is fine) and tol (0.1 is fine).

BFGS2 R R
Broyden-Fletcher-Goldfarb-Shanno (BFGS) algorithm, takes size of first trial step (0.1 is fine) and tol (0.1 is fine).

predictBinary :: Matrix -> Vector -> Vector

Binary Classification prediction function. Takes a matrix of features X and a vector theta. Returns predicted class.

predictMulti :: Matrix -> [Vector] -> Vector

Multiclass Classification prediction function. Takes a matrix of features X and a list of vectors theta, returns predicted class number assuming that class numbers start at 0.

calcAccuracy :: Vector -> Vector -> R

Calculates accuracy of Classification predictions. Takes vector expected y and vector predicted y. Returns number from 0 to 1, the closer to 1 the better accuracy. Suitable for both Classification Types: Binary and Multiclass.

processOutputMulti :: Int -> Vector -> [Vector]

Process outputs for Multiclass Classification. Takes number of labels and output vector y. Returns list of vectors of binary outputs (One-vs-All Classification). It is supposed that labels are integerets start at 0.

learnBinary

Arguments

:: MinimizeMethod
(e.g. BFGS2 0.1 0.1)

-> R
epsilon, desired precision of the solution;

-> Int
maximum number of iterations allowed;

-> R
regularization parameter lambda;

-> Matrix
matrix X;

-> Vector
binary vector y;

-> Vector
initial Theta;

-> (Vector, Matrix)
solution vector and optimization path.

Learns Binary Classification.

learnMulti

Arguments

:: MinimizeMethod
(e.g. BFGS2 0.1 0.1)

-> R
epsilon, desired precision of the solution;

-> Int
maximum number of iterations allowed;

-> R
regularization parameter lambda;

-> Matrix
matrix X;

-> [Vector]
list binary vector's y (one for every class, see processOutputMulti function);

-> [Vector]
initial theta;

-> ([Vector], [Matrix])
solution vector and optimization path.

Learns Multiclass Classification