MachineLearning.NeuralNetwork

Description

Simple Neural Networks.

Documentation

newtype NeuralNetworkModel

Neural Network Model. Takes neural network topology as a constructor argument.

Constructors

NeuralNetwork Topology

data Topology

Neural network topology has at least 2 elements: numver of input and number of outputs. And sizes of hidden layers between 2 elements. Bias input must not be included.

makeTopology :: Int -> Int -> [Int] -> Topology

Creates toplogy. Takes number of inputs, number of outputs and list of hidden layers.

initializeTheta :: Topology -> IO Vector

Create and initialize weights vector with random values for given neural network topology.

predictMulti :: Topology -> Matrix -> Vector -> Vector

Calculate output for given input matrix (X) and list of weight vector (theta). Returns vector of predicted class indices (assuming class indices starts at 0).

calcAccuracy :: Vector -> Vector -> R

Calculates accuracy of Classification predictions. Takes vector expected y and vector predicted y. Returns number from 0 to 1, the closer to 1 the better accuracy. Suitable for both Classification Types: Binary and Multiclass.

exported for testing purposes only

flatten :: [Matrix] -> Vector

Flatten list of matrices into vector.

unflatten :: Topology -> Vector -> [Matrix]

Unflatten vector into list of matrices for given neural network topology.

getThetaSizes :: Topology -> [(Int, Int)]

Returns dimensions of weight matrices for given neural network topology

initializeThetaList :: Topology -> IO [Matrix]

Create and initialize list of weights matrices with random values for given neural network topology.