MachineLearning.NeuralNetwork.Sigmoid

Description

Sigmoid

Documentation

makeTopology :: Int -> Int -> [Int] -> Topology

Creates toplogy. Takes number of inputs, number of outputs and list of hidden layers.

initializeTheta :: Int -> Topology -> Vector

Create and initialize weights vector with random values for given neural network topology. Takes a seed to initialize generator of random numbers as a first parameter.

initializeThetaIO :: Topology -> IO Vector

Create and initialize weights vector with random values for given neural network topology inside IO Monad.

initializeThetaM :: RandomGen g => Topology -> Rand g Vector

Create and initialize weights vector with random values for given neural network topology inside RandomMonad.