MachineLearning.NeuralNetwork.Topology

Description

Neural Network's Topology

Documentation

data Topology

Neural network topology has at least 2 elements: numver of input and number of outputs. And sizes of hidden layers between 2 elements.

type LossFunc = Matrix -> Matrix -> R

Loss function's type. Takes x, weights and y.

makeTopology :: Int -> [Layer] -> Layer -> LossFunc -> Topology

Makes Neural Network's Topology. Takes number of inputs, list of hidden layers, output layer and loss function.

loss :: Topology -> Regularization -> Matrix -> [(Matrix, Matrix)] -> Matrix -> R

Calculates loss for the given topology. Takes topology, regularization, x, weights, y.

propagateForward :: Topology -> Matrix -> [(Matrix, Matrix)] -> (Matrix, [Cache])

Implementation of forward propagation algorithm.

propagateBackward :: Topology -> Regularization -> Matrix -> [Cache] -> Matrix -> [(Matrix, Matrix)]

Implementation of backward propagation algorithm.

numberOutputs :: Topology -> Int

Returns number of outputs of the given topology.

initializeTheta :: Int -> Topology -> Vector

Create and initialize weights vector with random values for given neural network topology. Takes a seed to initialize generator of random numbers as a first parameter.

initializeThetaIO :: Topology -> IO Vector

Create and initialize weights vector with random values for given neural network topology inside IO Monad.

initializeThetaM :: RandomGen g => Topology -> Rand g Vector

Create and initialize weights vector with random values for given neural network topology inside RandomMonad.

flatten :: [(Matrix, Matrix)] -> Vector

Flatten list of matrices into vector.

unflatten :: Topology -> Vector -> [(Matrix, Matrix)]

Unflatten vector into list of matrices for given neural network topology.