MachineLearning.PCA

Description

Principal Component Analysis (PCA) - dimensionality reduction algorithm. It is mostly used to speed up supervising learning (Regression, Classification, etc) and visualization of data.

Documentation

getDimReducer :: Matrix -> Int -> (Matrix -> Matrix, R, Matrix)

Gets dimensionality reduction function, retained variance (0..1) and reduced X for given matrix X and number of dimensions to retain.

getDimReducer_rv :: Matrix -> R -> (Matrix -> Matrix, Int, Matrix)

Gets dimensionality reduction function, retained number of dimensions and reduced X for given matrix X and variance to retain (0..1].