MachineLearning.Random

Description

Randon generation uitility functions.

Documentation

sample :: RandomGen g => g -> Int -> Vector a -> (Vector a, g)

Samples n (given as a second parameter) values from list (given as a third parameter).

sampleM :: RandomGen g => Int -> Vector a -> Rand g (Vector a)

Samples n (given as a second parameter) values from list (given as a third parameter) inside RandomMonad.

getRandomRListM

Arguments

:: (RandomGen g, Random a)

=> Int
list's lengths

-> (a, a)
range

-> Rand g [a]
list of random values inside RandomMonad

Returns a list of random values distributed in a closed interval range

getRandomRVectorM

Arguments

:: RandomGen g

=> Int
vector's length

-> (R, R)
range

-> Rand g Vector

Returns a vector of random values distributed in a closed interval range

getRandomRMatrixM

Arguments

:: RandomGen g

=> Int
number of rows

-> Int
number of columns

-> (R, R)
range

-> Rand g Matrix

Returns a matrix of random values distributed in a closed interval range