MachineLearning.TerminalProgress

Description

Learn function with progress bar for terminal.

Documentation

learnWithProgressBar :: (Vector -> (Vector, Matrix)) -> Vector -> Int -> IO (Vector, Matrix)

Learn the given function displaying progress bar in terminal. It takes function, initial theta and number of iterations to call the function. It returns theta and optimization path (see MachineLearning.Optimization for details).

learnOneVsAllWithProgressBar :: (Vector -> Vector -> (Vector, Matrix)) -> Vector -> [Vector] -> Int -> IO ([Vector], [Matrix])

Learn the given function displaying progress bar in terminal. It takes function, list of outputs and list of initial thetas and number of iterations to call the function. It returns list of thetas and list of optimization paths (see MachineLearning.Optimization for details).