MachineLearning.NeuralNetwork

Description

Simple Neural Networks.

Documentation

class Model a where

Minimal complete definition

hypothesis, cost, gradient

Methods

hypothesis :: a -> Matrix -> Vector -> Vector

Hypothesis function, a.k.a. score function (for lassifition problem) Takes X (m x n) and theta (n x 1), returns y (m x 1).

cost :: a -> R -> Matrix -> Vector -> Vector -> R

Cost function J(Theta), a.k.a. loss function. It takes regularizarion parameter lambda, matrix X (m x n), vector y (m x 1) and vector theta (n x 1).

gradient :: a -> R -> Matrix -> Vector -> Vector -> Vector

Gradient function. It takes regularizarion parameter lambda, X (m x n), y (m x 1) and theta (n x 1). Returns vector of gradients (n x 1).

newtype NeuralNetworkModel

Neural Network Model. Takes neural network topology as a constructor argument.

Constructors

NeuralNetwork Topology

data Topology

Neural network topology has at least 2 elements: numver of input and number of outputs. And sizes of hidden layers between 2 elements. Bias input must not be included.

makeTopology :: Int -> Int -> [Int] -> Topology

Creates toplogy. Takes number of inputs, number of outputs and list of hidden layers.

initializeTheta :: Int -> Topology -> Vector

Create and initialize weights vector with random values for given neural network topology. Takes a seed to initialize generator of random numbers as a first parameter.

initializeThetaIO :: Topology -> IO Vector

Create and initialize weights vector with random values for given neural network topology inside IO Monad.

initializeThetaM :: RandomGen g => Topology -> Rand g Vector

Create and initialize weights vector with random values for given neural network topology inside RandomMonad.

calcAccuracy :: Vector -> Vector -> R

Calculates accuracy of Classification predictions. Takes vector expected y and vector predicted y. Returns number from 0 to 1, the closer to 1 the better accuracy. Suitable for both Classification Types: Binary and Multiclass.

Exported for testing purposes only.

flatten :: [Matrix] -> Vector

Flatten list of matrices into vector.

unflatten :: Topology -> Vector -> [Matrix]

Unflatten vector into list of matrices for given neural network topology.

getThetaSizes :: Topology -> [(Int, Int)]

Returns dimensions of weight matrices for given neural network topology

getThetaTotalSize :: Topology -> Int

Return sum of dimensions of weight matrices for given neural network topology.

initializeThetaListM :: RandomGen g => Topology -> Rand g [Matrix]

Create and initialize list of weights matrices with random values for given neural network topology.